Basic GUI
First approach to the program GUI. Basically, a notebook with two tabs:
one for searching and one for adding entries.
Removal shall happen in the search tab. Not sure about editing, maybe
I'll just keep the 'remove and reinsert' as an edit method(editing
shouldn't happen often since it's a dictionary).