Show all entries
A new button in the search frame allows to see the entire list of
entries in the dictionary.

In the controller module(Controller class): The method to get all
entries alphabetically oredered has been updated and uses now the
selected locale to compute the order.
The drawback is: The Controller class now needs an additional parameter,
the locale.

Refactored some methods names in the App class.